requirements-optional.txt
timm
rioxarray
eigenpip
cartopy
